+++ ref-exchange/release_notes.md
@@ -1,7 +1,8 @@
 # Release Notes
 
 ### Version 1.4.2
-1. Let guardians can call remove_exchange_fee_liquidity;
+1. Let both guardians and owner can remove exchange liquidity to owner inner account by remove_exchange_fee_liquidity;
+2. Let both guardians and owner can withdraw owner token to owner wallet by withdraw_owner_token;
 
 ### Version 1.4.1
 1. Introduce Stable-Swap-Pool;